[Reland] Unify get_block_size
